Spelling: Use one epoch on each chunk of a pass
On each pass of the training dataset, we used to run multiple
epochs on each chunk. This is problematic, because it over-fits
the last-chunk in the expence of others.
Since what we want is essentially to emulate batched processing
we only need one epoch per chunk - so our our passes are the epochs.